requirements.txt
flask
requests
python-dotenv
gunicorn
flask_sqlalchemy
flask_cors
